Refactor : supabase 관련 파일을 utils 디렉토리에 위치
