+++ web/src/app/api/magazines/[id]/route.ts
@@ -2,7 +2,7 @@ import path from 'path'
 
 import { NextRequest, NextResponse } from 'next/server'
 
-import { supabaseServerClient } from '@/shared/lib'
+import { supabaseServerClient } from '@/shared/utils/supabase/server'
 
 export async function GET(
   request: NextRequest,
