+++ web/src/middleware.ts
@@ -1,6 +1,6 @@
 import { NextResponse, type NextRequest } from 'next/server'
 
-import { supabaseServerClient } from '@/shared/lib'
+import { supabaseServerClient } from '@/shared/utils/supabase/server'
 
 export async function middleware(request: NextRequest) {
   try {
